requirements.txt
requests>=2.32.3
lxml>=5.3.1
beautifulsoup4>=4.13.3
pandas>=2.2.3
google-cloud-storage>=3.1.0
google-cloud-firestore>=2.20.1
google-cloud-core>=2.4.3
tqdm>=4.67.1
secedgar>=0.5.0
arelle>=2.2